chore: move jinja2 templates to a separate package
