+++ projects/pgai/pgai/semantic_catalog/templates/__init__.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+from jinja2 import Environment, FileSystemLoader
+
+env = Environment(loader=FileSystemLoader(Path(__file__).parent.joinpath("templates")))
