+++ projects/pgai/pgai/semantic_catalog/templates/templates/procedure.j2
