+++ projects/pgai/pgai/semantic_catalog/templates/templates/sql_example.j2
