+++ projects/pgai/pgai/semantic_catalog/templates/templates/table.j2
