+++ projects/pgai/pgai/semantic_catalog/templates/templates/view.j2
