+++ projects/pgai/tests/semantic_catalog/test_describe.py
@@ -1,5 +1,6 @@
 import io
 import os
+from pathlib import Path
 
 import psycopg
 import pytest
@@ -7,8 +8,15 @@ import pytest
 import pgai
 import pgai.semantic_catalog as semantic_catalog
 import pgai.semantic_catalog.describe as describe
+from pgai.semantic_catalog.models import ObjectDescription
 
-from .utils import PostgresContainer, load_airports
+from .utils import (
+    PostgresContainer,
+    get_procedures,
+    get_tables,
+    get_views,
+    load_airports,
+)
 
 DATABASE = "sc_02"
 
@@ -210,7 +218,7 @@ async def test_find_procs(container: PostgresContainer):
 
 
 async def test_describe(container: PostgresContainer):
-    if "ANTHROPTIC_API_KEY" not in os.environ:
+    if "ANTHROPIC_API_KEY" not in os.environ:
         pytest.skip("No anthroptic api key provided")
     expected = {
         "postgres_air.airport",
@@ -267,4 +275,76 @@ async def test_describe(container: PostgresContainer):
             where description is not null
         """)
         actual = {row[0] for row in await cur.fetchall()}
+    assert actual == expected
+
+
+async def test_render_description_to_sql(container: PostgresContainer):
+    actual = Path(__file__).parent.joinpath("data", "render_description_to_sql.actual")
+    with actual.open("w") as f:
+        async with await psycopg.AsyncConnection.connect(
+            container.connection_string(database="postgres_air")
+        ) as con:
+            for i, table in enumerate(get_tables()):
+                desc = ObjectDescription(
+                    classid=42,
+                    objid=i,
+                    objsubid=0,
+                    objtype="table",
+                    objnames=[table.schema_name, table.table_name],
+                    objargs=[],
+                    description=f"this is a description for table {table.table_name}",
+                )
+                f.write(describe.render_sql(con, "my_catalog", desc))
+                if table.columns:
+                    for col in table.columns:
+                        desc = ObjectDescription(
+                            classid=42,
+                            objid=i,
+                            objsubid=col.objsubid,
+                            objtype="table column",
+                            objnames=[table.schema_name, table.table_name, col.name],
+                            objargs=[],
+                            description=f"this is a description for column {col.name}",
+                        )
+                        f.write(describe.render_sql(con, "my_catalog", desc))
+            for i, view in enumerate(get_views()):
+                desc = ObjectDescription(
+                    classid=42,
+                    objid=i + 100,
+                    objsubid=0,
+                    objtype="view",
+                    objnames=[view.schema_name, view.view_name],
+                    objargs=[],
+                    description=f"this is a description for view {view.view_name}",
+                )
+                f.write(describe.render_sql(con, "my_catalog", desc))
+                if view.columns:
+                    for col in view.columns:
+                        desc = ObjectDescription(
+                            classid=42,
+                            objid=i + 100,
+                            objsubid=col.objsubid,
+                            objtype="view column",
+                            objnames=[view.schema_name, view.view_name, col.name],
+                            objargs=[],
+                            description=f"this is a description for column {col.name}",
+                        )
+                        f.write(describe.render_sql(con, "my_catalog", desc))
+            for i, procedure in enumerate(get_procedures()):
+                desc = ObjectDescription(
+                    classid=666,
+                    objid=i,
+                    objsubid=0,
+                    objtype=procedure.kind,
+                    objnames=[procedure.schema_name, procedure.proc_name],
+                    objargs=procedure.objargs,
+                    description=f"this is a description for {procedure.kind} {procedure.proc_name}",  # noqa: E501
+                )
+                f.write(describe.render_sql(con, "my_catalog", desc))
+    actual = actual.read_text()
+    expected = (
+        Path(__file__)
+        .parent.joinpath("data", "render_description_to_sql.expected")
+        .read_text()
+    )
     assert actual == expected
